feat(task-53-grid-expansion-64): Expand pattern grid to 64 steps
Core changes:
- Update kMaxSteps from 32 to 64 in DuoPulseTypes.h
- Update kMaxPhraseSteps from 256 to 512
- Add kMaxSubSteps = 256 for flam resolution support

Type migration (uint32_t -> uint64_t):
- Pattern masks in PatternGenerator.h (anchorMask, shimmerMask, auxMask)
- VoiceRelation functions (FindGaps, ShiftMask*, etc.)
- GuardRails functions (SoftRepairPass, ApplyHardGuardRails, etc.)
- GumbelSampler selection functions (SelectHitsGumbelTopK, etc.)
- HitBudget eligibility masks and constants
- EuclideanGen pattern generation functions
- DriftControl stability mask
- VelocityCompute accent masks

Bit operations updated:
- All (1U << step) changed to (1ULL << step)
- Loop bounds updated from 32 to 64 where appropriate
- Mask constants expanded to 64-bit patterns

Tests updated:
- 28 test files modified for 64-bit compatibility
- All 373 tests pass (62942 assertions)

Memory impact:
- FLASH: 120,724 B (92.11% of 128 KB) - minimal increase
- SRAM: 30,456 B (5.81% of 512 KB) - unchanged

+++ src/Engine/DriftControl.h
@@ -78,7 +78,7 @@ float GetStepStability(int step, int patternLength);
  * @param stabilityThreshold Only include steps with stability >= this
  * @return Bitmask of stable steps
  */
-uint32_t GetStabilityMask(int patternLength, float stabilityThreshold);
+uint64_t GetStabilityMask(int patternLength, float stabilityThreshold);
 
 // =============================================================================
 // Seed Selection Functions
